Updating Android backend
