+++ Windows/NatReader.h
@@ -26,24 +26,17 @@
 
 
 #pragma region --MediaReader--
-
 /*!
- @function NRCreateMP4FrameReader
+ @function NRDispose
  
- @abstract Create an MP4 frame reader.
+ @abstract Dispose a media reader.
  
- @discussion Create an MP4 frame reader.
+ @discussion Dispose a media reader.
  
- @param uri
- URL to media source. MUST be prepended with URI scheme/protocol.
- 
- @param startTime
- Media time to start reading samples in seconds.
- 
- @param duration
- Duration in seconds.
+ @param reader
+ Opaque pointer to a media reader.
  */
-BRIDGE EXPORT void* APIENTRY NRCreateMP4FrameReader (const char* uri, float startTime, float duration);
+BRIDGE EXPORT void APIENTRY NRDisposeReader (void* reader);
 
 /*!
  @function NRMediaURI
@@ -71,55 +64,21 @@ BRIDGE EXPORT void APIENTRY NRMediaURI (void* reader, char* dstString);
  Opaque handle to a media reader.
  */
 BRIDGE EXPORT float APIENTRY NRMediaDuration (void* reader);
-
-/*!
- @function NRCopyNextFrame
- 
- @abstract Copy the next frame from the reader.
- 
- @discussion Copy the next frame from the reader.
- 
- @param reader
- Opaque handle to a media reader.
- 
- @param buffer
- Destination buffer.
- 
- @param outBufferSize
- Number of bytes copied in bytes.
- 
- @param outTimestamp
- Timestamp of the copied buffer.
- */
-BRIDGE EXPORT void APIENTRY NRCopyNextFrame (void* reader, void* buffer, int32_t* outBufferSize, int64_t* outTimestamp);
-
-/*!
- @function NRReset
- 
- @abstract Reset the reader to start reading from the initial posiiton.
- 
- @discussion Reset the reader to start reading from the initial posiiton.
- 
- @param reader
- Opaque handle to a media reader.
- */
-BRIDGE EXPORT void APIENTRY NRReset (void* reader);
-
-/*!
- @function NRDispose
- 
- @abstract Dispose a media reader.
- 
- @discussion Dispose a media reader.
- 
- @param reader
- Opaque pointer to a media reader.
- */
-BRIDGE EXPORT void APIENTRY NRDispose (void* reader);
 #pragma endregion
 
 
 #pragma region --FrameReader--
+/*!
+ @function NRCreateMP4FrameReader
+ 
+ @abstract Create an MP4 frame reader.
+ 
+ @discussion Create an MP4 frame reader.
+ 
+ @param uri
+ URL to media source. MUST be prepended with URI scheme/protocol.
+ */
+BRIDGE EXPORT void* APIENTRY NRCreateMP4FrameReader (const char* uri);
 
 /*!
  @function NRFrameSize
@@ -150,4 +109,55 @@ BRIDGE EXPORT void APIENTRY NRFrameSize (void* frameReader, int32_t* outWidth, i
  Opaque handle to a frame reader.
  */
 BRIDGE EXPORT float APIENTRY NRFrameRate (void* frameReader);
+#pragma endregion
+
+
+#pragma region --MediaEnumerator--
+/*!
+ @function NRCreateEnumerator
+ 
+ @abstract Create a media frame enumerator to read frames from the reader.
+ 
+ @discussion Create a media frame enumerator to read frames from the reader.
+ 
+ @param startTime
+ Media time to start reading samples in seconds.
+    
+ @param duration
+ Duration in seconds.
+ */
+BRIDGE EXPORT void* APIENTRY NRCreateEnumerator (void* reader, float startTime, float duration);
+
+/*!
+ @function NRDisposeEnumerator
+ 
+ @abstract Dispose a media enumerator.
+ 
+ @discussion Dispose a media enumerator.
+ 
+ @param enumerator
+ Opaque handle to a media enumerator.
+ */
+BRIDGE EXPORT void APIENTRY NRDisposeEnumerator (void* enumerator);
+
+/*!
+ @function NRCopyNextFrame
+ 
+ @abstract Copy the next frame from the reader.
+ 
+ @discussion Copy the next frame from the reader.
+ 
+ @param enumerator
+ Opaque handle to a media enumeratorr.
+ 
+ @param dstBuffer
+ Destination buffer.
+ 
+ @param outBufferSize
+ Number of bytes copied in bytes.
+ 
+ @param outTimestamp
+ Timestamp of the copied buffer.
+ */
+BRIDGE EXPORT void APIENTRY NRCopyNextFrame (void* enumerator, void* dstBuffer, int32_t* outBufferSize, int64_t* outTimestamp);
 #pragma endregion
